start piano eval script refactor
